`textformat` is a utility module. Move it.
