+++ salt/utils/textformat.py
