+++ salt/utils/thin.py
@@ -789,7 +789,7 @@ def gen_min(cachedir, extra_mods='', overwrite=False, so_mods='',
         'salt/minion.py',
         'salt/pillar',
         'salt/pillar/__init__.py',
-        'salt/textformat.py',
+        'salt/utils/textformat.py',
         'salt/log',
         'salt/log/__init__.py',
         'salt/log/handlers',
